% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blastn_taxo_assingment.R
\name{blast_assignment}
\alias{blast_assignment}
\title{Loads blast assignment}
\usage{
blast_assignment(
  blastapp_path,
  queries,
  megablast_opts = "-evalue 0.001 -max_target_seqs 5 -perc_identity 0.8",
  blastn_opts = "-evalue 0.001 -max_target_seqs 5 -perc_identity 0.5",
  db,
  output_path,
  nthreads,
  minSim = 97,
  minCov = 80,
  update = TRUE,
  pident = "no",
  taxonly = "TRUE"
)
}
\description{
This function takes in a fasta file and returns taxonomic assignments.
Specifically, it performs blastn and megablast in the nt database of NCBI and
then uses a Last Common Ancestor (LCA) approach and a minimum percent identity
(per taxonomic rank) to assign taxonomy. It then merges the results from blastn
and megablast using the highest taxonomic resolution with a consensus between
the two approaches.
}
\examples{
blast_assignment(blastapp_path = "/home/miniconda3/bin/blastn",
queries="uniqueSeqs.fasta",
megablast_opts="-evalue 0.001 -max_target_seqs 5 -perc_identity 0.8",
blastn_opts="-evalue 0.001 -max_target_seqs 5 -perc_identity 0.5",db="nt",
output="blast_resutls",
nthreads=20)
}
