% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_micro.R
\name{core_micro}
\alias{core_micro}
\title{core_micro}
\usage{
core_micro(pst, grp, method = NULL, n = NULL, d = NULL, p = NULL, ...)
}
\arguments{
\item{pst}{Phyloseq object}

\item{grp}{Group to perform the core microbiome subsetting by.}

\item{method}{Subsetting methodology. Two methods are currently available: 'core' and 'crab'}

\item{n}{Divider value for the 'crab' method. To keep ASVs with cumulative relative abundance within two orders of magnitude of the maximum cumulative relative abundance for example, n would be set at 100.}

\item{d}{Detection threshold for the 'core' function.}

\item{p}{Prevalence threshold for the 'core' function.}

\item{...}{Other arguments to pass to the 'core' function.}
}
\description{
This function takes in a phyloseq object and subset core taxa per a specified group as opposed to subsetting by the overall data.
Two methods  currently available are 1) the 'core' function of the microbiome R package (Lahti et al., 2019) and 2) subsetting by Cumulative Relative ABundance (crab).
The latter is defined by setting a divider to the maximum cumulative abundance, which will be used as a minimum threshold value to keep taxa.
If you use this function for a publication, please cite the creators of the microbiome R package (citation('microbiome')).
}
\examples{
core_micro(pst = ps_test_data,  grp = "biome", method = "crab", n = 10)
}
