% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_taxo_assignment.R
\name{extra_taxo_assignment}
\alias{extra_taxo_assignment}
\title{extra_taxo_assignment}
\usage{
extra_taxo_assignment(obj, sqlFile)
}
\arguments{
\item{obj}{Object}

\item{sqlFile}{Path to the local NCBI taxonomy db}
}
\description{
This function takes in a pyloseq object with taxonomy data or a dataframe containing taxonomy information and currently adds the Protozoa and
Archaeplastida (excl. Viridiplantae) groups under the Kingdom rank. This script is particularly useful
if taxonomy has been normalised using NCBI curated classification and nomenclature database
(using the taxo_normalisation function of this package for example) where protozoans and Archaeplastida (excl. Viridiplantae)
are generally not assigned any information under Kingdom, making it difficult to dissociate metazoans from protists, or the Plantae group.
Importantly, this function requires the download of NCBI taxonomic database.
This takes few minutes to install using the following command:\cr
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package. The database is ~65 GB but the user can set getAccessions=FALSE to drastically reduce its size.
}
\examples{
data("ps_test_data")
extra_taxo_assignment(obj = ps_test_data, sqlFile = 'accessionTaxa.sql')
}
