% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginext_t3_bc.R
\name{gginext_t3_bc}
\alias{gginext_t3_bc}
\title{gginext_t3_bc}
\usage{
gginext_t3_bc(inext)
}
\arguments{
\item{inext}{iNEXT onject}
}
\description{
This function takes in an iNEXT object and create a ggiNEXT plot of type 3 (Diversity per sample coverage) that contains a vertical line at base coverage (defined as the highest coverage value between minimum extrapolated values and maximum interpolated values)
}
\examples{
inext_test = inext_ifreq_wrapper(pst = ps_test_data,  grp = "biome", q = 0, knots = 100, nboot = 200)
gginext_t3_bc(inext = inext_test)
}
