% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/improved_tax_legend.R
\name{improved_tax_legend}
\alias{improved_tax_legend}
\title{improved_tax_legend}
\usage{
improved_tax_legend(
  p,
  rank1 = "Phylum",
  rank2 = "Family",
  others_last = FALSE,
  legend_position = "bottom"
)
}
\arguments{
\item{p}{Plot from the taxo_bar_plot function of biohelper}

\item{rank1}{Higher taxonomic rank associated with colours (e.g. Phylum)}

\item{rank2}{Lower taxonomic rank associated with shades of colours (e.g. Family)}

\item{others_last}{Optionally places the groups "Others" last, within each higher taxonomic group}

\item{legend_position}{Position of the legend}
}
\description{
This function takes in a plot from the taxo_bar_plot function of biohelper and returns the figure with a modified legend, where the higher rank (e.g. Phylum \link{rank1}) is displayed in bold only once
above its associated lower groups (e.g. Families \link{rank2}).
The function is simply meant to provide a cleaner legend for this type of figure.
}
\examples{
ps_test_data_t = ps_test_data \%>\% tax_glom('Family') \%>\% subset_samples(extraction_method!="na") \%>\% microbiome::transform(transform = "compositional")\cr
p1 = taxo_bar_plot(ps_test_data_t, rank1 = "Phylum", rank2 = "Family") + facet_wrap(extraction_method~., drop = TRUE, scale="free", nrow = 1) + ggtitle("Taxonomic composition per extraction method")\cr
improved_tax_legend(p1,rank1 = "phylum", rank2 = "family")

}
