% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inext_ifreq_wrapper.R
\name{inext_ifreq_wrapper}
\alias{inext_ifreq_wrapper}
\title{inext_ifreq_wrapper}
\usage{
inext_ifreq_wrapper(pst, grp, ...)
}
\arguments{
\item{pst}{Phyloseq object}

\item{grp}{Group to compare}

\item{...}{Other arguments to pass to the 'iNEXT' function.}
}
\description{
This function takes in a phyloseq object and perform iNEXT from the iNEXT R package (Hsieh et al., 2022) on incidence frequency data (Chao2 method).
Briefly, the data is split into a list of dataframes based on the grouping variable and frequency of incidence per group is calculated.
}
\examples{
inext_ifreq_wrapper(pst = ps_test_data,  grp = "biome", q = 0, knots = 100, nboot = 200)
}
