% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs_to_Xs.R
\name{NAs_to_Xs}
\alias{NAs_to_Xs}
\title{NAs_to_Xs}
\usage{
NAs_to_Xs(df)
}
\arguments{
\item{df}{Taxonomic table whose missing values should be replaced.}
}
\description{
This function replaces NA entries in a taxonomic table by the previous (higher)
taxonomic information and adds Xs according to the number of NA entries for each organism.
}
\details{
The first column must contain ASV/OTU identifiers as in the example below:\tabular{llll}{
   ASV \tab Kingdom \tab Phylum \tab Genus \cr
   ASV_1 \tab Eukaryota \tab NA \tab NA \cr
   ASV_2 \tab Eukaryota \tab Chordata \tab Mus \cr
   ASV_3 \tab Eukaryota \tab Chordata \tab NA \cr
}
}
\examples{
NAs_to_Xs(ps_test_data@tax_table@.Data \%>\% as.data.frame())
}
