% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalized_otu_taxo.R
\name{normalized_otu_taxo}
\alias{normalized_otu_taxo}
\title{Loads normalized_otu_taxo}
\usage{
normalized_otu_taxo(physeq_obj, otu_column = "otu", asv_column = "asv")
}
\arguments{
\item{physeq_obj}{= A phyloseq object}

\item{otu_column}{= Name of the OTU column}

\item{asv_column}{= Name of the ASV column}
}
\description{
This function normalises taxonomic assignments of ASVs belonging to the same OTU by using
the taxonomic assignment of the ASV that has the highest taxonomic resolution.
}
\examples{
# Normalizing taxonomy followed by merging ASVs to OTUs
normalized_otu_taxo(physeq_obj = ps_obj) \%>\% speedyseq::tax_glom(taxrank = "otu")


}
