% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_decon.R
\name{ps_decon}
\alias{ps_decon}
\title{ps_decon}
\usage{
ps_decon(
  ps_temp,
  method = "complete_asv_removal",
  group = NA,
  runs = 2,
  thresh = 0.7,
  prop.thresh = 5e-05,
  regression = 0,
  low.threshold = 40,
  up.threshold = 400,
  min_reads = 10
)
}
\arguments{
\item{ps_temp}{Phyloseq object to decontaminate.}

\item{method}{Method to be used for decontamination. Options are
\code{"microDecon"} (using \code{\link[microDecon:decon]{microDecon::decon()}}), \code{"max_v"} and
\code{"complete_asv_removal"}.}

\item{group}{Metadata column used in the \code{numb.ind} argument of
\code{\link[microDecon:decon]{microDecon::decon()}}.}

\item{runs, thresh, prop.thresh, regression, low.threshold, up.threshold}{Arguments passed to \code{\link[microDecon:decon]{microDecon::decon()}} when \code{method = "microDecon"}.}

\item{min_reads}{Minimum sample read count retained before running
\code{method = "microDecon"}.}
}
\description{
This function offers 3 different options to remove putative sequence contamination
of a phyloseq object. The microDecon option is a wrapper of the decon function
from microDecon R package. It performs the decon function on a phyloseq object
with sample data and returns a decontaminated phyloseq object with sample data,
taxonomy and reference sequences if present.

The 'max_v' option subtract read associated to putative contaminant ASV by using
their max read count in blank(s).

Finally, the 'complete_asv_removal' removes all ASVs found in blanks from the dataset.

Note that the 'max_v' option does not take into account the compositional nature of the data.

Also note that microDecon assumes a common source of contamination. Quoting the authors,
"If substantial differences among blanks occur only across experimental blocks, such as
extraction kits (suggesting consistent contamination within a block), then use microDecon
separately for each block. If, however, there is substantial variability among blanks within
blocks (suggesting contamination from poor laboratory techniques), microDecon will not be effective.

The sample_data must have a column labeled sample_id and a column labeled amplicon_type, and the
non blank samples must be labeled as 'sample' (case-insensitive). Alternatively, the user can provide a list of samples
to use as controls.

If you use the decon function for publication, please cite the authors of the microDecon R package (citation('microDecon')).

For example:\tabular{lllll}{
   sample_id \tab amplicon_type \tab extraction_batch \tab extraction_method \tab etc. \cr
   sample1 \tab sample \tab 1 \tab manual \tab NA \cr
   sample2 \tab sampling_blank \tab 1 \tab manual \tab NA \cr
   sample3 \tab dna_extraction_blank \tab 2 \tab robot \tab NA \cr
   sample4 \tab sample \tab 2 \tab robot \tab NA \cr
   sample5 \tab pcr_blank \tab 2 \tab robot \tab NA \cr
}
}
\examples{
ps_decon(ps_test_data, method = "microDecon", group = "dna_extraction_batch_id")
}
