% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ps_test_data}
\alias{ps_test_data}
\title{Example phyloseq object}
\format{
A \code{phyloseq} object.
}
\source{
Included with \code{biohelper} for package examples.
}
\usage{
ps_test_data
}
\description{
A small \code{phyloseq} object used in examples and tests.
}
\keyword{datasets}
