% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psto_inext_ifreq.R
\name{psto_inext_ifreq}
\alias{psto_inext_ifreq}
\title{psto_inext_ifreq}
\usage{
psto_inext_ifreq(pst, q = 0, grp, ...)
}
\arguments{
\item{pst}{Phyloseq object}

\item{q}{A number or vector specifying the diversity order(s) of Hill numbers.}

\item{grp}{Group to compare}

\item{...}{Other arguments to pass to the 'iNEXT' function.}
}
\description{
This function takes in a phyloseq object and perform iNEXT from the iNEXT R package (Hsieh et al., 2022) on incidence frequency data (Chao2 method).
Briefly, the data is split into a list of dataframes based on the grouping variable and frequency of incidence per group is calculated.
If you use this function for a publication, please cite the creators of the iNEXT R package (citation('iNEXT')).
}
\examples{
inext_ifreq_wrapper(pst = ps_test_data,  grp = "biome", q = 0, knots = 100, nboot = 200)
}
