% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psto_upset.R
\name{psto_upset}
\alias{psto_upset}
\title{psto_upset}
\usage{
psto_upset(
  pst,
  grp,
  order.by = "freq",
  sets.x.label = "Taxa richness",
  nrow = NULL,
  align = NULL,
  rel_heights = NULL,
  rel_widths = NULL,
  ...
)
}
\arguments{
\item{pst}{Phyloseq object}

\item{grp}{Group to compare shared and unique taxon.}

\item{order.by}{How the intersections in the matrix should be ordered by. Options include frequency (entered as "freq"), degree, or both in any order.}

\item{sets.x.label}{The x-axis label of the set size bar plot.}

\item{rel_heights}{Optional argument for the plot_grid function: Numerical vector of relative rows heights. Works just as rel_widths does, but for rows rather than columns.}

\item{rel_widths}{Optional argument for the plot_grid function: Numerical vector of relative columns widths. For example, in a two-column grid, rel_widths = c(2, 1) would make the first column twice as wide as the second column.}

\item{...}{Other arguments to pass to the 'upset' function.}
}
\description{
This function is a wrapper to create an 'upset' figure from a phyloseq object and using the UpSetR package by Gehlenborg (2019).
It returns a dataframe for UpsetR, the plot obtained directly from the 'upset' function and a ggplot2 object obtained from the cowplot::plot_grid function for convenience.
If you use this function for a publication, please cite the creators of the UpSetR and cowplot R packages (citation('UpSetR') and citation('cowplot')).
}
\examples{
psto_upset(pst = ps_test_data, grp = "biome")
}
