% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstoveg_otu.R
\name{pstoveg_otu}
\alias{pstoveg_otu}
\title{pstoveg_otu}
\usage{
pstoveg_otu(PS)
}
\arguments{
\item{PS}{(required) a phyloseq object}
}
\description{
This function from the theseus R package takes in a phyloseq object and returns a OTU/ASV table compatible with the vegan R package.
If you use this function for a publication, please cite the creators of the theseus R package (citation('theseus')).
}
\examples{
pstoveg_otu(PS = ps_test_data)
}
