% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pstoveg_sample.R
\name{pstoveg_sample}
\alias{pstoveg_sample}
\title{pstoveg_sample}
\usage{
pstoveg_sample(PS)
}
\arguments{
\item{PS}{(required) a phyloseq object}
}
\description{
This function from the theseus R package takes in a phyloseq object and returns a metadata table compatible with the vegan R package.
If you use this function for a publication, please cite the creators of the theseus R package (citation('theseus')).
}
\examples{
pstoveg_sample(PS = ps_test_data)
}
