% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_group_dissimilarity.R
\name{pw_group_dissimilarity}
\alias{pw_group_dissimilarity}
\title{Pairwise dissimilarity per group.}
\usage{
pw_group_dissimilarity(
  physeq,
  group = NULL,
  between_groups = TRUE,
  method = "bray",
  method_title = FALSE,
  notch = TRUE,
  justDF = FALSE,
  ...
)
}
\arguments{
\item{physeq}{A phyloseq-class object}

\item{group}{Grouping variable name (contained in \code{\link{sample_data}})}

\item{between_groups}{Logical, estimate between-group dissimilarity}

\item{method}{Distance/dissimilarity method (as character string; see \code{\link{distanceMethodList}})}

\item{method_title}{Logical, add method name to the plot title}

\item{notch}{Logical, draw the notches at each side of the boxes}

\item{justDF}{Logical, instead of returning a ggplot2-object, return the data.frame that was used to build the plot}

\item{...}{Additional arguments may be passed to \code{\link{distance}} function from phyloseq package}
}
\value{
ggplot2-object or data.frame (if justDF = TRUE).
}
\description{
Pairwise dissimilarity per group.
}
\details{
This script is an adaptation of the 'phyloseq_group_dissimilarity' function of the metagMisc R package to include jaccard nestedness (jne) and turnover (jtu) as dissimilarity methods. If you use it, please cite the original package (citation('metagMisc')).
If the notches of two boxplots do not overlap this indicates that the two medians differ.
}
\examples{
## Load and subset data
data(enterotype)
ent <- subset_samples(enterotype, Enterotype \%in\% c("1", "2"))

## Dissimilarity boxplots
phyloseq_group_dissimilarity(ent, group = "Enterotype")
phyloseq_group_dissimilarity(ent, group = "Enterotype", between_groups = F)
phyloseq_group_dissimilarity(ent, group = "Enterotype", method = "jaccard")
}
