% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_decon.R
\name{simple_decon}
\alias{simple_decon}
\title{Loads simple_decon}
\usage{
simple_decon(ps_obj, method = "complete_asv_removal")
}
\description{
This function takes in a phyloseq object with sample data and performs a simple decontamination of ASV data
by either completely removing contaminant ASVs and their associated reads across the dataset ("complete_asv_removal")
or by subtracting read count per ASV using their max read count in blank(s) ("max_v"). The latter option does not
take into account the compositional nature of the data. If using the latter option,
the user may want to perform this step per batch of sample (e.g. DNA extraction or PCR batch).
The sample_data MUST have a column labeled sample_id and a column labeled amplicon_type.
Non blank samples must be labeled "sample".
For example:
sample_id   amplicon_type.        DNA_extraction_batch      extraction_method   etc.
sample1     sample                1                         manual
sample2     sampling_blank        1                         manual
sample3     dna_extraction_blank  2                         robot
sample4     sample                2                         robot
sample5     pcr_blank             2                         robot
}
\examples{
simple_decon(ps_test_data, method = "complete_asv_removal")
simple_decon(ps_test_data, method = "max_v")
}
