% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_assignment_summary.R
\name{taxo_assignment_summary}
\alias{taxo_assignment_summary}
\title{taxo_assingment_summary}
\usage{
taxo_assignment_summary(df, ranks)
}
\arguments{
\item{df}{Dataframe of taxonomic assignment}

\item{ranks}{Taxonomic ranks in the dataframe}
}
\description{
This function takes in a dataframe of taxonomic assignment and returns the mean number and standard deviation of taxonomic ranks assigned.
}
\examples{
df = ps_test_data@tax_table \%>\% as.data.frame()
taxo_assingment_summary(df = df, ranks = c("Kingdom","Phylum","CLass","Order","Family","Genus"))
}
