% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_merge.R
\name{taxo_merge}
\alias{taxo_merge}
\title{taxo_merge}
\usage{
taxo_merge(df_list, sqlFile, ranks, keepSAR = F)
}
\arguments{
\item{sqlFile}{Path to NCBI taxonomic reference database}

\item{ranks}{Ranks to return}

\item{keepSAR}{Keep the SAR assignment from the input data, which is not a valid group in NCBI taxonomy db (default is FALSE)}

\item{df1}{First dataframe with taxonomic assignment}

\item{df2}{Second dataframe with taxonomic assignment}
}
\description{
This function takes in a list of taxonomic assignment dataframes, normalises the
taxonomy using NCBI curated taxonomic database, and uses the highest
taxonomic resolution between the assignments if there is consensus across
all assigned ranks and dataframes. Otherwise, it assigns taxonomy the Last common
ancestor (LCA) if there is a consensus (>50\%) across dataframes.

Importantly, this function requires the download of NCBI taxonomic database (~65 GB).
Additionally, the first column of the dataframe needs to be the feature id (e.g. ASV, OTU, MOTU, ZOTU identifier, etc.).

This takes few minutes to install using the following command:
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package.
}
\examples{
taxo_merge(df_list = list(df1, df2),  sqlFile = path_to_NCBI_taxo_db, ranks = c("Superkingdom","Kingdom","Phylum","Class","Order","Family","Genus","Species"), keepSAR=F)
}
