% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_psmelt.R
\name{tidy_psmelt}
\alias{tidy_psmelt}
\title{Loads tidy_psmelt}
\usage{
tidy_psmelt(physeq)
}
\description{
The tidy_psmelt function is a specialized melt function for melting phyloseq objects (instances of the phyloseq class), usually for producing graphics with ggplot2.
This function was originally created and posted by mikemc (https://github.com/mikemc) to solve an issue with phyloseq psmelt function (https://github.com/joey711/phyloseq/issues/1035)
}
\examples{

tidy_psmelt(ps_test_data)
}
