% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_to_venn.R
\name{upset_to_venn}
\alias{upset_to_venn}
\title{upset_to_venn}
\usage{
upset_to_venn(
  upset_df,
  set_labels = NULL,
  palette = NULL,
  alpha = 0.75,
  legend = TRUE,
  quant_cex = 0.75,
  set_cex = 1,
  two_line = FALSE,
  min_pct_to_label = 0,
  percent_digits = 1,
  bold_sets = TRUE
)
}
\arguments{
\item{upset_df}{Dataframe from the list returned by psto_upset. The dataframe should contain groups as columns and variables as row, using a presence (1) / absence (0) format.}

\item{set_labels}{Labels for groups}

\item{palette}{Palette of colors to use}

\item{alpha}{Transparency level}

\item{legend}{Whether to include a legend or not}

\item{quant_cex}{Quantity text size}

\item{set_cex}{Size of set names}

\item{two_line}{Whether to display quantity and proportion on the same line or not}

\item{min_pct_to_label}{Hide labels below this \% of union (e.g., 1.0)}

\item{percent_digits}{Number of percent digits to use}

\item{bold_sets}{Set labels in bold}
}
\description{
This function takes in the dataframe created from the psto_upset function and outputs a proportional venn diagram using the
eulerr R package.
If you use this function for a publication, please cite the creators of the UpSetR and eulerr R packages (citation('UpSetR') and citation('eulerr')).
}
\examples{
upset_list = psto_upset(pst = ps_test_data, grp = "extraction_method")
upset_to_venn(upset_df = upset_list$upset_df)

}
