% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Negate_func.R
\name{\%ni\%}
\alias{\%ni\%}
\title{Load a \%ni\%}
\usage{
... \%ni\% NA
}
\description{
This function performs a Negate operation similar to \%in\%.
}
\examples{
str_a = c("A","B","C")
"A" \%ni\% str_a
"D" \%ni\% str_a
}
