% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcaPident.R
\name{lcaPident}
\alias{lcaPident}
\title{Loads lcaPident}
\usage{
lcaPident(
  blastn_file,
  output,
  minSim = 97,
  minCov = 80,
  update = FALSE,
  pident = "no",
  pgenus = 95,
  pfamily = 87,
  porder = 83,
  pclass = 81,
  pphylum = 79,
  pkingdom = 71,
  taxonly = "TRUE"
)
}
\arguments{
\item{blastn_file}{Blast file containing multiple hits per query sequence. To work with this function the blastn search MUST have the format output 6  with options "query.id", "query.length", "pident", "subject.id", "subject.GBid", "evalue", "bit.score","staxids", "sscinames", "sblastnames", "qcovs", "qcovhsp"
minSim             Minimum similarity to assign species hits (default: 97)
minCov             Minimum coverage to keep hits (default: 80)
update             Should the taxonomy database be updated? (default: FALSE)
pident             To reduce taxonomy assingment according to default percent identity thresholds. Options are: before or after LCA assingment
pgenus             Minimum similarity to assign genus (default: 95)
pfamily            Minimum similarity to assign family (default: 87)
porder             Minimum similarity to assign order (default: 83)
pclass             Minimum similarity to assign class (default: 81)
pphylum            Minimum similarity to assign phylum (default: 79)
pkingdom           Minimum similarity to assign kingdom (default: 71)}
}
\description{
This  function takes in the output of a blastn search with multiple hits and
uses a Last Common Ancestor (LCA) approach and optionally, a minimum percent
identity (pident; per taxonomic rank) to assign taxonomy.
}
\examples{
lcaPident(blastn_file = blastn_file, output = "taxo_assignment.csv", pident="no")
}
