% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NAs_to_Xs.R
\name{NAs_to_Xs}
\alias{NAs_to_Xs}
\title{Loads NAs_to_Xs}
\usage{
NAs_to_Xs(df)
}
\description{
This function replaces NA entries in a taxonomic table by the previous (higher)
taxonomic information and adds Xs according to the number of NA entries for each organism.
}
\details{
The first column must contain ASV/OTU identifiers as the example below:
ASV     Kingdom     Phylum   Genus
ASV_1   Eukaryota
ASV_2   Eukaryota   Chordata  Mus
ASV_3   Eukaryota   Chordata
}
\examples{
NAs_to_Xs(ps_test_data@tax_table@.Data \%>\% as.data.frame())
}
