% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_decon.R
\name{ps_decon}
\alias{ps_decon}
\title{ps_decon}
\usage{
ps_decon(
  ps_temp,
  method = "complete_asv_removal",
  group = NA,
  runs = 2,
  thresh = 0.7,
  prop.thresh = 5e-05,
  regression = 0,
  low.threshold = 40,
  up.threshold = 400,
  min_reads = 10
)
}
\arguments{
\item{method}{Method to be used for decontamination. Options are 'microDecon' (using the decon function of microDecon), 'max_v' and 'complete_asv_removal'\cr}

\item{group}{Will be used in the numb.ind argument of the microDecon::decon function\cr}

\item{ps}{Phyloseq object to decontaminate\cr}

\item{(...)}{If using microDecon the user can specify any argument of the decon function with the exception of num.blanks and numb.ind, which are already handled by ps_decon.\cr}
}
\description{
This function offers 3 different options to remove putative sequence contamination
of a phyloseq object. The microDecon option is a wrapper of the decon function
from microDecon R package. It performs the decon function on a phyloseq object
with sample data and returns a decontaminated phyloseq object with sample data,
taxonomy and reference sequences if present.

The 'max_v' option subtract read associated to putative contaminant ASV by using
their max read count in blank(s).

Finally, the 'complete_asv_removal' removes all ASVs found in blanks from the dataset.

Note that the 'max_v' option does not take into account the compositional nature of the data.

Also note that microDecon assumes a common source of contamination. Quoting the authors,
"If substantial differences among blanks occur only across experimental blocks, such as
extraction kits (suggesting consistent contamination within a block), then use microDecon
separately for each block. If, however, there is substantial variability among blanks within
blocks (suggesting contamination from poor laboratory techniques), microDecon will not be effective.

The sample_data must have a column labeled sample_id and a column labeled amplicon_type, and the
non blank samples must be labeled as 'sample'. Alternatively, the user can provide a list of samples
to use as controls.

If you use the decon function for publication, please cite the authors of the microDecon R package (citation('microDecon')).

For example:\cr
\tabular{rrrrr}{
\strong{sample_id} \tab \strong{amplicon_type} \tab \strong{extraction_batch} \tab \strong{extraction_method} \tab \strong{etc.} \cr
sample1 \tab sample \tab 1 \tab manual \tab NA\cr
sample2 \tab sampling_blank \tab 1 \tab manual \tab NA\cr
sample3 \tab dna_extraction_blank \tab 2 \tab  robot \tab NA\cr
sample4 \tab sample \tab 2 \tab robot \tab NA\cr
sample5 \tab pcr_blank \tab 2 \tab robot \tab NA
}
}
\examples{
ps_decon(ps_test_data, method = "microDecon", group = "dna_extraction_batch_id")
}
