% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_to_microDecon.R
\name{ps_to_microDecon}
\alias{ps_to_microDecon}
\title{Loads ps_to_microDecon}
\usage{
ps_to_microDecon(
  ps,
  groups = NA,
  runs = 2,
  thresh = 0.7,
  prop.thresh = 5e-05,
  regression = 0,
  low.threshold = 40,
  up.threshold = 400
)
}
\description{
This function is a wrapper of the decon function from microDecon R package.
It performs the decon funtion on a phyloseq object with sample data and
returns a decontaminated phyloseq object with sample data, taxonomy and reference sequences if present.
The sample_data MUST have a column labeled sample_id and a column labeled sample_type.
Non blank samples must be labeled 'sample'
For example:
sample_id   blank_type            DNA_extraction_batch      extraction_method   etc.
sample1     sample                1                         manual
sample2     sampling_blank        1                         manual
sample3     dna_extraction_blank  2                         robot
sample4     sample                2                         robot
sample5     pcr_blank             2                         robot
}
\examples{
ps_to_microDecon(ps_test_data, groups = "extraction_method")
}
