% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_bar_plot.R
\name{taxo_bar_plot}
\alias{taxo_bar_plot}
\title{Loads taxo_bar_plot}
\usage{
taxo_bar_plot(
  ps_obj,
  rank1 = "Phylum",
  rank2 = "Family",
  n_rank1 = NA,
  n_rank2 = 6,
  x_labels = NA,
  colors = c("cyan", "palegreen", "yellow", "deeppink ", "white", "dodgerblue",
    "lightsalmon"),
  f_ord = NA
)
}
\description{
This function takes in a phyloseq object with taxonomy and returns a bar plot with colours and shades per
taxa. The output is not directly customizable but can be re-used to add facets for example by adding factors of interest (f).
}
\details{
rank1 = Taxonomic rank to be associated with colours (e.g. Phylum)
rank2 = Taxonomic rank to be associated with shades of colours (e.g. Family)
colors = Colors to be used
n_rank2 = Number of shades to use per colour. Taxa which rank below that number will be assigned to "Other".
f_ord = Order of the samples in the plot using sample name
}
\examples{
colors = c("cyan", "palegreen", "yellow", "deeppink ", "white", "dodgerblue", "lightsalmon")
ps_test_data_t = ps_test_data \%>\% tax_glom('Family') \%>\% microbiome::transform(transform = "compositional")
p1 = taxo_bar_plot(ps_test_data_t, rank1 = "Phylum", rank2 = "Family", colors = colors,  f = "extraction_method")
p1 + facet_wrap(extraction_method~., drop = TRUE, scale="free", nrow = 1) + ggtitle("Taxonomic composition per extraction method")
}
