% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_merge_old_version.R
\name{taxo_merge}
\alias{taxo_merge}
\title{taxo_merge}
\usage{
taxo_merge(df1, df2, sqlFile, ranks, keepSAR = F)
}
\arguments{
\item{df1}{First dataframe with taxonomic assignment}

\item{df2}{Second dataframe with taxonomic assignment}

\item{sqlFile}{Path to NCBI taxonomic reference database}

\item{ranks}{Ranks to return}

\item{keepSAR}{Keep the SAR assignment from the input data, which is not a valid group in NCBI taxonomy db (default is FALSE)}
}
\description{
This function takes in two taxonomic assignment dataframes, normalises the
taxonomy using NCBI curated taxonomic database, and uses the highest
taxonomic resolution between the two assignments if there is consensus across
all assigned ranks. Otherwise, it assigns taxonomy using a Last common
ancestor approach (LCA).

Importantly, this function requires the download of NCBI taxonomic database (~65 GB).

This takes few minutes to install using the following command:
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package.
}
\examples{
taxo_merge(df1 = df1, df2 = df2,  sqlFile = path_to_NCBI_taxo_db, ranks = c("Superkingdom","Kingdom","Phylum","Class","Order","Family","Genus","Species"), keepSAR=F)
}
