% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_normalisation.R
\name{taxo_normalisation}
\alias{taxo_normalisation}
\title{Loads taxo_normalisation}
\usage{
taxo_normalisation(obj, sqlFile, ranks, keepSAR = F)
}
\arguments{
\item{obj}{= Either a dataframe containing taxonomic information (e.g. output of Qiime2 or the assignTaxonomy from dada2) or a phyloseq object}

\item{sqlFile}{= Path to the local NCBI taxonomy db}

\item{ranks}{= Ranks to return}

\item{keepSAR}{= Keep the SAR assignment from the input data, which is not a valid group in NCBI taxonomy db (default is FALSE)}
}
\description{
This function performs normalisation of taxonomic assignments of 18S and COI data
identified via silva, midori or any other reference databases, using NCBI curated
classification, nomenclature and database. It can also provide further taxonomic
information of additional desired ranks. It can either take in a taxonomy table
(e.g. from Qiime2) or a phyloseq object.

Importantly, this function requires the download of NCBI taxonomic database. 

This takes few minutes to install using the following command:\cr
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package. The database is relatively large (~65GB) but the size can be drastically reduced by setting getAccessions=FALSE
}
\examples{
data("ps_test_data")
taxo_normalisation(ps_test_data, sqlFile = 'accessionTaxa.sql', ranks = c("Superkingdom", "Kingdom", "Phylum",  "Class",   "Order",   "Family",  "Genus"))

}
