% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_normalisation.R
\name{taxo_normalisation}
\alias{taxo_normalisation}
\title{Load a taxo_normalisation}
\usage{
taxo_normalisation(df, sqlFile, ranks)
}
\description{
This function performs normalisation of taxonomic assignments of 18S and COI data
identified via silva, midori or any other reference databases, using NCBI curated
classification and nomenclatureand database. It can also provide further taxonomic
information of additional desired ranks.
}
\details{
The first column must contain ASV/OTU identifiers as the example below:
ASV     Kingdom   Phylum   etc
ASV_1   Eukaryota   Arthropoda
ASV_2   Eukaryota   Chordata
ASV_3   Eukaryota   Chordata

Importantly, this function requires the download of NCBI taxonomic database (~65 GB).
This takes few minutes to install using the following command:
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package
}
\examples{
taxo_normalisation(taxo_dataframe, sqlFile = sqlFile, desired_ranks = c("Family", "Genus", "Species"))
}
