% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_normalisation.R
\name{taxo_normalisation}
\alias{taxo_normalisation}
\title{Loads taxo_normalisation}
\usage{
taxo_normalisation(obj, sqlFile, ranks, keepSAR = F)
}
\arguments{
\item{obj}{= Either a dataframe containing taxonomic information (e.g. output of Qiime2 or the assignTaxonomy from dada2) or a phyloseq object}

\item{sqlFile}{= Path to the local NCBI taxonomy db}

\item{ranks}{= Ranks to return}

\item{keepSAR}{= Keep the SAR assignment from the input data, which is not a valid group in NCBI taxonomy db. If false, SAR taxa are kept but the 'SAR' assignation is removed. Note that some databases may not use the 'SAR' label so if merging data assigned with different databases and KeepSAR is TRUE, there may be discrepancies in the data (default is FALSE)}
}
\description{
This function performs normalisation of taxonomic assignments of 18S and COI data
identified via silva, midori or any other reference databases, using NCBI curated
classification and nomenclatureand database. It can also provide further taxonomic
information of additional desired ranks. It can either take in a taxonomy table
(e.g. from Qiime2) or a phyloseq object.

Importantly, this function requires the download of NCBI taxonomic database.

This takes few minutes to install using the following command:\cr
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package. The database is ~65 GB but the user can set getAccessions=FALSE to drastically reduce its size.
}
\examples{
data("ps_test_data")
taxo_normalisation(ps_test_data, sqlFile = 'accessionTaxa.sql', ranks = c("Superkingdom", "Kingdom", "Phylum",  "Class",   "Order",   "Family",  "Genus"))

}
