% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxo_normalisation.R
\name{taxo_normalisation}
\alias{taxo_normalisation}
\title{Loads taxo_normalisation}
\usage{
taxo_normalisation(obj, sqlFile, ranks, keepSAR = F)
}
\description{
This function performs normalisation of taxonomic assignments of 18S and COI data
identified via silva, midori or any other reference databases, using NCBI curated
classification and nomenclatureand database. It can also provide further taxonomic
information of additional desired ranks. It can either take in a taxonomy table
(e.g. from Qiime2) or a phyloseq object.
}
\details{
Importantly, this function requires the download of NCBI taxonomic database (~65 GB).
This takes few minutes to install using the following command:
prepareDatabase('accessionTaxa.sql') # From the taxonomizr R package
}
\examples{
data("ps_test_data")
taxo_normalisation(ps_test_data, sqlFile = 'accessionTaxa.sql', ranks = c("Superkingdom", "Kingdom", "Phylum",  "Class",   "Order",   "Family",  "Genus"))

}
